moved sign out button to header
