+++ src/components/posts/PostsPage.js
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 import PostForm from './PostForm';
 import client from '../../utils/client';
 import './style.css';
@@ -10,7 +9,6 @@ const PostsPage = () => {
   const [post, setPost] = useState({ content: '' });
   const [postResponse, setPostResponse] = useState('');
   const [posts, setPosts] = useState([]);
-  let navigate = useNavigate();
 
   useEffect(() => {
     client.get('/posts').then((res) => setPosts(res.data.data.posts));
@@ -35,19 +33,10 @@ const PostsPage = () => {
     });
   };
 
-  const signOut = (event) => {
-    event.preventDefault();
-    localStorage.setItem(process.env.REACT_APP_USER_TOKEN, '');
-    navigate('../', { replace: true });
-  };
-
   return (
     <>
       <Header />
       <section className='posts-section'>
-        <button id='user-signout-button' onClick={signOut}>
-          sign out
-        </button>
         <p>Status: {postResponse.status}</p>
         <PostForm handleSubmit={createPost} handleChange={handleChange} />
         <ul className='posts-list'>
